Connection part on hook, new HUD
