+++ src/components/HUD/HUD.module.scss
@@ -1,15 +1,52 @@
 .hud {
-  margin-top: 10px;
-  text-align: center;
-  justify-self: center;
-  max-width: 400px;
-  width: 100%;
-  color: white;
-  background-color: rgba(0, 0, 0, 0.5);
-  padding: 20px;
-  border-radius: 5px;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+  align-items: center;
+
+  > div {
+    text-align: center;
+    max-width: 400px;
+    width: 100%;
+    color: white;
+    background-color: rgba(0, 0, 0, 0.5);
+    padding: 20px;
+    border-radius: 5px;
+  }
 
   h1 {
     margin: 0;
   }
+}
+
+.player-list {
+  .header {
+    display: flex;
+    justify-content: space-between;
+  }
+
+  .content {
+    width: 100%;
+    text-align: left;
+  }
+}
+
+.controls {
+}
+
+.copied-message {
+  position: absolute;
+  top: 50%;
+  left: 50%;
+  transform: translate(-50%, -50%);
+  background-color: rgba(0, 0, 0, 0.7);
+  color: white;
+  padding: 10px;
+  border-radius: 5px;
+  opacity: 1;
+  transition: opacity 0.5s ease-in-out;
+}
+
+.copied-message.fade-out {
+  opacity: 0;
 }
